duplicateReferences folder: Remove describe clause "payloads with duplicate references" and rename files
